Revert "fix steamtypes.h"
This reverts commit 87823c4680c6bcd87be4a0a18520b63f56cd9e97.

+++ public/steam/steamtypes.h
@@ -80,8 +80,8 @@ typedef unsigned long int ulint64;
 typedef int64 intp;
 typedef uint64 uintp;
 #else
-typedef int32 intp;
-typedef uint32 uintp;
+typedef int64 intp;
+typedef uint64 uintp;
 #endif
 
 #endif // else _WIN32
